# Improved Memory Efficiency for `ppanggolin fasta` Command