The `ppanggolin fasta` command could use less memory. Right now, it loads a lot of data to build the pangenome object, which makes filtering sequences easy but comes with a big memory and time cost, especially for large pangenomes with thousands of genomes.

This PR changes the command to read directly from the HDF5 tables and write the sequences on the fly, reducing the load. Some intermediate tables are only loaded when needed.

The arguments `--genes`, `--proteins`, `--prot_families`, and `--gene_families` have been optimized. Only `--regions` works the same as before, as it needs more work to optimize.