@JeanMainguy 2024-09-20
Benchmark on an E. coli pangenome made of 5k genomes, resulting in a 2.4 GB HDF5 file. 

| command                   | PeakMemory (GB) dev | PeakMemory (GB) master | time dev (min) | time master (min) |
|---------------------------|---------------------|------------------------|----------------|-------------------|
| genes_softcore             | 5.2                 | 37.4                   | 9.0            | 26.2              |
| genes_core                 | 4.8                 | 36.3                   | 3.4            | 24.1              |
| gene_families_softcore     | 4.8                 | 36.3                   | 3.3            | 24.1              |
| genes_all                  | 3.0                 | 33.6                   | 9.5            | 25.4              |
| prot_families_core         | 4.8                 | 32.9                   | 1.4            | 20.6              |
| genes_persistent           | 3.8                 | 32.5                   | 7.5            | 25.3              |
| genes_rgp                  | 2.4                 | 32.1                   | 4.2            | 25.1              |
| genes_shell                | 2.6                 | 31.9                   | 5.1            | 24.4              |
| genes_cloud                | 1.2                 | 31.4                   | 2.3            | 23.3              |
| gene_families_cloud        | 1.1                 | 31.3                   | 2.0            | 23.1              |
| gene_families_all          | 1.1                 | 31.3                   | 1.6            | 23.4              |
| gene_families_shell        | 1.1                 | 31.3                   | 1.5            | 23.2              |
| genes_module_0             | 1.1                 | 31.3                   | 2.1            | 22.5              |
| genes_module_1             | 1.1                 | 31.3                   | 1.6            | 23.1              |
| gene_families_module_1     | 1.1                 | 31.3                   | 1.4            | 23.1              |
| gene_families_module_0     | 1.1                 | 31.3                   | 1.4            | 23.1              |
| proteins_module_1          | 1.1                 | 31.3                   | 2.0            | 24.2              |
| proteins_module_0          | 1.1                 | 31.3                   | 2.0            | 23.4              |
| prot_families_rgp          | 1.2                 | 28.4                   | 0.4            | 21.5              |
| prot_families_module_0     | 0.5                 | 11.1                   | 0.1            | 5.2               |
| prot_families_module_1     | 0.5                 | 11.1                   | 0.1            | 5.2               |
| prot_families_persistent   | 0.5                 | 11.1                   | 0.1            | 5.2               |
| prot_families_shell        | 0.5                 | 11.1                   | 0.1            | 5.2               |
| prot_families_all          | 0.5                 | 11.1                   | 0.1            | 5.3               |